Add files via upload
Together with Luca Malferrari and Filippo Ugolini, we developed this project as part of the final examination for the Optimal Control course in the Automation Engineering program at the University of Bologna.

The goal of this project is to design an optimal control strategy for an actuated flexible surface,
which models the behavior of adaptive wing structures in aircraft or unmanned aerial vehicles
(UAVs). 

For further details, please refer to the attached PDF report.

Best regards,
Enrico Battistini.